Move code to a shared place that can be imported from both
